requirements.txt
django
gunicorn
django-heroku
google-api-python-client
google-cloud
google-cloud-vision